onCaseInputCreate cloud function
