+++ functions/src/screen.ts
@@ -14,25 +14,6 @@ export const  onSearchCreate = functions.firestore
 		await userTriggeredScreen(document, document.data()!['target'], 2, 0.9, userId);
 	});
 
-// onCaseInputCreate creates a document inside 'input' collection when a user types into the input field on CASES page
-
-export const onCaseInputCreate = functions.firestore
-	.document("user/{userId}/case/{caseId}/input/{inputId}")
-	.onWrite(async (change, context) => {
-	//let userId = document.ref.path.split("/")[1]
-		const oldData = change.before.data()
-		const newData = change.after.data()
-
-		// Triggered on UI Input creation/updation in the text field
-		const newInput = newData?.text;
-		const oldInput = oldData?.text;
-
-		if(newInput !== oldInput){
-			console.log(`OLD: ${oldInput},  NEW: ${newInput}`)
-		}
-
-})
-
 export const  onCaseSearchCreate = functions.firestore
 	.document("user/{userId}/case/{caseId}/search/{searchId}")
 	.onCreate(async (document: DocumentSnapshot, context) => {
